Break the upgrade command into component files like other commands
